requirements.txt
streamlit>=1.35
pandas>=2.0
numpy>=1.24
altair>=5.0
pyyaml>=6.0
